requirements.txt
pytesseract >= 0.3.7
Pillow~=8.2.0
pdf2image~=1.15.1